+++ src/components/calcite-value-list-item/calcite-value-list-item.tsx
@@ -54,7 +54,8 @@ export class CalciteValueListItem {
   @Prop({ mutable: true }) handleActivated? = false;
 
   /**
-   * Determines the icon SVG symbol that will be shown. Options are circle, square, grid or null.
+   * Determines the icon SVG symbol that will be shown. Options are circle, square, grip or null.
+   * @see [ICON_TYPES](https://github.com/Esri/calcite-components/blob/master/src/components/calcite-pick-list/resources.ts#L5)
    */
   @Prop({ reflect: true }) icon?: ICON_TYPES | null = null;